+++ dart/dynamics/MetaSkeleton.cpp
@@ -30,10 +30,12 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include "dart/dynamics/MetaSkeleton.hpp"
+
+#include <algorithm>
 #include "dart/common/Console.hpp"
 #include "dart/dynamics/DegreeOfFreedom.hpp"
 #include "dart/dynamics/JacobianNode.hpp"
-#include "dart/dynamics/MetaSkeleton.hpp"
 
 namespace dart {
 namespace dynamics {