+++ dart/dynamics/Skeleton.cpp
@@ -382,6 +382,13 @@ std::mutex& Skeleton::getMutex() const
 }
 
 //==============================================================================
+std::unique_ptr<common::LockableReference> Skeleton::getLockableReference() const
+{
+  return common::make_unique<common::SingleLockableReference<std::mutex>>(
+      mPtr, mMutex);
+}
+
+//==============================================================================
 Skeleton::~Skeleton()
 {
   for (BodyNode* bn : mSkelCache.mBodyNodes)